requirements.txt
annotated-types==0.6.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-genai==1.10.0
googleapis-common-protos==1.63.0
grpcio # ==1.63.0
grpcio-status # ==1.62.2
httplib2 # ==0.22.0
idna # ==3.7
itsdangerous # ==2.2.0
Jinja2 # ==3.1.5
MarkupSafe # ==2.1.5
pillow # ==11.0.0
proto-plus # ==1.23.0
protobuf # ==4.25.3
pyasn1 # ==0.6.0
pyasn1_modules # ==0.4.0

pydantic # >=2.8.0

# pydantic==2.7.1
pydantic_core  #==2.18.2
pyparsing # ==3.1.2
python-dotenv # ==1.0.1
requests # ==2.32.2
rsa # ==4.9
tqdm #==4.66.4
typing_extensions #==4.11.0
uritemplate #==4.1.1
urllib3 #==2.2.2
Werkzeug #==3.0.6